+++ hahomematic/platforms/decorators.py
@@ -10,7 +10,7 @@ from functools import wraps
 import logging
 from typing import Any, ParamSpec, TypeVar
 
-import hahomematic
+from hahomematic.context import IN_SERVICE_VAR
 from hahomematic.exceptions import BaseHomematicException
 from hahomematic.support import reduce_args
 
@@ -141,17 +141,17 @@ def service(log_level: int = logging.ERROR) -> Callable:
         async def service_wrapper(*args: P.args, **kwargs: P.kwargs) -> T:
             """Wrap service to log exception."""
             token: Token | None = None
-            if not hahomematic.IN_SERVICE_VAR.get():
-                token = hahomematic.IN_SERVICE_VAR.set(True)
+            if not IN_SERVICE_VAR.get():
+                token = IN_SERVICE_VAR.set(True)
             try:
                 return_value = await func(*args, **kwargs)
                 if token:
-                    hahomematic.IN_SERVICE_VAR.reset(token)
+                    IN_SERVICE_VAR.reset(token)
                 return return_value  # noqa: TRY300
             except BaseHomematicException as bhe:
                 if token:
-                    hahomematic.IN_SERVICE_VAR.reset(token)
-                if not hahomematic.IN_SERVICE_VAR.get() and log_level > logging.NOTSET:
+                    IN_SERVICE_VAR.reset(token)
+                if not IN_SERVICE_VAR.get() and log_level > logging.NOTSET:
                     logging.getLogger(args[0].__module__).log(
                         level=log_level, msg=reduce_args(args=bhe.args)
                     )